pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repodoctor"
version = "0.1.0"
description = "CLI tool that detects technical debt in codebases"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "RepoDoctor" }]
keywords = ["technical-debt", "code-analysis", "cli", "static-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.24.0",
    "tree-sitter-javascript>=0.24.0",
    "tree-sitter-typescript>=0.21.0",
    "gitpython>=3.1.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
repodoctor = "repodoctor.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
